fix(aicc-test): isolate judge model selection
